requirements.txt
google-genai
Pillow
python-dotenv
requests
Flask
boto3
pillow_heif